refactor: split server/mod.rs into graph, enrichment, and dispatch
Move graph lifecycle (get_graph, build_full_graph, update_graph_with_scan)
to server/graph.rs and background enrichment (LSP, embedding, scanner) to
server/enrichment.rs. mod.rs retains RnaHandler struct, MCP dispatch, and
tests. Pure code movement — no logic changes.

Closes #254
[outcome:context-assembly]
